+++ src/com/intellij/idea/plugin/hybris/project/DefaultHybrisProjectImportBuilder.java
@@ -1,6 +1,7 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019 EPAM Systems <hybrisideaplugin@epam.com>
+ * Copyright (C) 2014-2016 Alexander Bartash <AlexanderBartash@gmail.com>
+ * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -25,7 +26,6 @@ import com.intellij.idea.plugin.hybris.notifications.Notifications;
 import com.intellij.idea.plugin.hybris.project.configurators.ConfiguratorFactory;
 import com.intellij.idea.plugin.hybris.project.configurators.PostImportConfigurator;
 import com.intellij.idea.plugin.hybris.project.descriptors.*;
-import com.intellij.idea.plugin.hybris.project.descriptors.impl.AbstractYModuleDescriptor;
 import com.intellij.idea.plugin.hybris.project.descriptors.impl.RootModuleDescriptor;
 import com.intellij.idea.plugin.hybris.project.tasks.ImportProjectProgressModalWindow;
 import com.intellij.idea.plugin.hybris.project.tasks.SearchModulesRootsTaskModalWindow;
@@ -85,7 +85,7 @@ public class DefaultHybrisProjectImportBuilder extends AbstractHybrisProjectImpo
     public void setRootProjectDirectory(@NotNull final File directory) {
         Validate.notNull(directory);
 
-        LOG.info("setting RootProjectDirectory to " + directory.getAbsolutePath());
+        LOG.info("setting RootProjectDirectory to "+directory.getAbsolutePath());
         ProgressManager.getInstance().run(new SearchModulesRootsTaskModalWindow(
             directory, this.getHybrisProjectDescriptor()
         ));
@@ -180,8 +180,8 @@ public class DefaultHybrisProjectImportBuilder extends AbstractHybrisProjectImpo
             : message("hybris.notification.project.import.title");
 
         Notifications.create(NotificationType.INFORMATION, notificationTitle,
-                             message("hybris.notification.import.or.refresh.process.not.finished.yet.content"))
-                     .notify(project);
+                message("hybris.notification.import.or.refresh.process.not.finished.yet.content"))
+            .notify(project);
     }
 
     protected void performProjectsCleanup(@NotNull final Iterable<ModuleDescriptor> modulesChosenForImport) {
@@ -238,13 +238,13 @@ public class DefaultHybrisProjectImportBuilder extends AbstractHybrisProjectImpo
     @Override
     public List<ModuleDescriptor> getBestMatchingExtensionsToImport(final @Nullable HybrisProjectSettings settings) {
         final List<ModuleDescriptor> allModules = this.getHybrisProjectDescriptor().getFoundModules();
-        final Set<ModuleDescriptor> moduleToImport = new LinkedHashSet<>();
+        final List<ModuleDescriptor> moduleToImport = new ArrayList<>();
         final Set<ModuleDescriptor> moduleToCheck = new HashSet<>();
         for (ModuleDescriptor moduleDescriptor : allModules) {
             if (moduleDescriptor.isPreselected()) {
+                moduleToImport.add(moduleDescriptor);
                 moduleDescriptor.setImportStatus(MANDATORY);
                 moduleToCheck.add(moduleDescriptor);
-                resolveSubModules(moduleDescriptor, moduleToImport, MANDATORY, moduleToCheck);
             }
         }
         resolveDependency(moduleToImport, moduleToCheck, MANDATORY);
@@ -256,9 +256,9 @@ public class DefaultHybrisProjectImportBuilder extends AbstractHybrisProjectImpo
         allModules.stream()
             .filter(e -> unusedExtensionNameSet.contains(e.getName()))
             .forEach(e -> {
+                moduleToImport.add(e);
                 e.setImportStatus(UNUSED);
                 moduleToCheck.add(e);
-                resolveSubModules(e, moduleToImport, UNUSED, moduleToCheck);
             });
         resolveDependency(moduleToImport, moduleToCheck, UNUSED);
 
@@ -267,27 +267,27 @@ public class DefaultHybrisProjectImportBuilder extends AbstractHybrisProjectImpo
             : Collections.emptySet();
 
         return moduleToImport.stream()
-                             .filter(e -> !modulesOnBlackList.contains(e.getRelativePath()))
-                             .sorted(Comparator.nullsLast(Comparator.comparing(ModuleDescriptor::getName)))
-                             .collect(Collectors.toList());
+            .filter(e -> !modulesOnBlackList.contains(e.getRelativePath()))
+            .sorted(Comparator.nullsLast(Comparator.comparing(ModuleDescriptor::getName)))
+            .collect(Collectors.toList());
     }
 
     @Override
     public void setCoreStepModuleList() {
         moduleList = this.getHybrisProjectDescriptor()
-                         .getFoundModules()
-                         .stream()
-                         .filter(Predicate.not(RootModuleDescriptor.class::isInstance))
-                         .collect(Collectors.toList());
+            .getFoundModules()
+            .stream()
+            .filter(Predicate.not(RootModuleDescriptor.class::isInstance))
+            .collect(Collectors.toList());
     }
 
     @Override
     public void setExternalStepModuleList() {
         moduleList = this.getHybrisProjectDescriptor()
-                         .getFoundModules()
-                         .stream()
-                         .filter(RootModuleDescriptor.class::isInstance)
-                         .collect(Collectors.toList());
+            .getFoundModules()
+            .stream()
+            .filter(RootModuleDescriptor.class::isInstance)
+            .collect(Collectors.toList());
     }
 
     @Override
@@ -335,7 +335,7 @@ public class DefaultHybrisProjectImportBuilder extends AbstractHybrisProjectImpo
     }
 
     private void resolveDependency(
-        final Set<ModuleDescriptor> moduleToImport,
+        final List<ModuleDescriptor> moduleToImport,
         final Set<ModuleDescriptor> moduleToCheck,
         final ModuleDescriptorImportStatus selectionMode
     ) {
@@ -347,7 +347,6 @@ public class DefaultHybrisProjectImportBuilder extends AbstractHybrisProjectImpo
                         moduleToImport.add(moduleDescriptor);
                         moduleDescriptor.setImportStatus(selectionMode);
                         moduleToCheck.add(moduleDescriptor);
-                        resolveSubModules(moduleDescriptor, moduleToImport, MANDATORY, moduleToCheck);
                     }
                 }
             }
@@ -355,16 +354,4 @@ public class DefaultHybrisProjectImportBuilder extends AbstractHybrisProjectImpo
         }
     }
 
-    private static void resolveSubModules(final ModuleDescriptor moduleDescriptor, final Set<ModuleDescriptor> moduleToImport, final ModuleDescriptorImportStatus mandatory, final Set<ModuleDescriptor> moduleToCheck) {
-        if (moduleDescriptor instanceof final AbstractYModuleDescriptor yModule) {
-            for (final var subModule : yModule.getSubModules()) {
-                if (moduleToImport.contains(subModule)) {
-                    continue;
-                }
-                moduleToImport.add(subModule);
-                subModule.setImportStatus(mandatory);
-                moduleToCheck.add(subModule);
-            }
-        }
-    }
 }