+++ CHANGELOG.md
@@ -54,7 +54,7 @@
 - Added `hmc` sub-module dependency on `hmc.web` [#574](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/574) 
 - Improved detection of the Gradle modules [#575](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/575) 
 - Use main config folder during project import [#576](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/576) 
-- Added submodules to be imported to modules.xml dependencies [#554](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/554)
+- Fixed header text for Other modules step of the Project Import Wizard [#580](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/580) 
 
 ### Other
 - Adjusted inline documentation for Type System [#539](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/539)